correctly match the kqueue man page

+++ src/sys/event.rs
@@ -15,7 +15,7 @@ mod ffi {
         pub flags: EventFlag,       // 2
         pub fflags: FilterFlag,     // 4
         pub data: intptr_t,         // 8
-        pub udata: u64              // 8
+        pub udata: uint             // 8
     }
 
     // Bug in rustc, cannot determine that kevent is #[repr(C)]
@@ -163,7 +163,7 @@ pub fn ev_set(ev: &mut KEvent,
               filter: EventFilter,
               flags: EventFlag,
               fflags: FilterFlag,
-              udata: u64) {
+              udata: uint) {
 
     ev.ident  = ident as u64;
     ev.filter = filter;
